refactor: File, Path가 존재하지 않을 경우 처리 (멘토 피드백3)  - FileXXXRepository에서 Path, File이 존재하지 않을 경우에 대한 예외처리 작

+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileChannelRepository.java
@@ -58,12 +58,22 @@ public class FileChannelRepository implements ChannelRepository {
     @Override
     public void delete(Channel channel) {
         Path path = Path.of(directoryPath.toAbsolutePath() + "/" + channel.getId() + FileUtil.getExtension());
+
+        if(!path.toFile().exists()){
+            throw new IllegalArgumentException("존재하지 않는 채널입니다.");
+        }
+
         path.toFile().delete();
     }
 
     @Override
     public void deleteAll() {
         File directory = new File(directoryPath.toAbsolutePath() + "/");
+
+        if(!directory.exists() || !directory.isDirectory()){
+            throw new IllegalArgumentException("존재하지 않는 채널 디렉토리입니다.");
+        }
+
 
         File[] files = directory.listFiles();
         if(files != null){
