+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileMessageRepository.java
@@ -58,12 +58,21 @@ public class FileMessageRepository implements MessageRepository {
     @Override
     public void delete(Message message) {
         Path path = Path.of(directoryPath.toAbsolutePath() + "/" + message.getId() + FileUtil.getExtension());
+
+        if(!path.toFile().exists()){
+            throw new IllegalArgumentException("존재하지 않는 메시지입니다.");
+        }
+
         path.toFile().delete();
     }
 
     @Override
     public void deleteAll() {
         File directory = new File(directoryPath.toAbsolutePath() + "/");
+
+        if(!directory.exists() || !directory.isDirectory()){
+            throw new IllegalArgumentException("존재하지 않는 메시지 디렉토리입니다.");
+        }
 
         File[] files = directory.listFiles();
         if(files != null){
