+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileUserRepository.java
@@ -56,12 +56,21 @@ public class FileUserRepository implements UserRepository {
     @Override
     public void delete(User user) {
         Path path = Path.of(directoryPath.toAbsolutePath() + "/" + user.getId() + FileUtil.getExtension());
+
+        if(!path.toFile().exists()){
+            throw new IllegalArgumentException("존재하지 않는 유저입니다.");
+        }
+
         path.toFile().delete();
     }
 
     @Override
     public void deleteAll() {
         File directory = new File(directoryPath.toAbsolutePath() + "/");
+
+        if(!directory.exists() || !directory.isDirectory()){
+            throw new IllegalArgumentException("존재하지 않는 유저 디렉토리입니다.");
+        }
 
         File[] files = directory.listFiles();
         if(files != null){
